Solve 1st part of day 20
I started with TDD, but had some issues with understanding the flow.
Took a break and rewritten everything kinda similar to what I had,
but with inspiration taken from Redux/Flux and in a hacky way, w/o tests
as it was enough to confirm the final count to ensure it works.